+++ CHANGES.rst
@@ -14,12 +14,21 @@ New features:
 
 Bug fixes:
 
+- Fix migrating Archetypes references.  The previous store and restore
+  references approach was stuffing all references into the `relatesTo`
+  relationship because it uses the `relatedItems` AT field by default.  The
+  new approach migrates references per-instance migrated and adds a field
+  migrator for migrating `ReferenceField` fields into
+  `plone.app.relationfield` fields.
+  [rpatterson]
+
 - Fix migrating False boolean values, handling of "unset" values. If an AT
   schema had a `BooleanField` that defaults to `True` and an instance had a
   value set to `False`, the migration would take that `False` to mean to leave
   the value as unset on the new DX instance which would then get the default
   `True` value.  Fix this bug by properly handling unset values by checking
   for `AttributeError`.
+  [rpatterson]
 
 - Fix folder layout property migration. The default listing_view layout was
   always set if a folder didn't have a layout property (with older content